Separate event parts
