+++ pkg/ans/ans.go
@@ -7,22 +7,7 @@ import (
 	"github.com/SAP/jenkins-library/pkg/body"
 	"github.com/SAP/jenkins-library/pkg/xsuaa"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	"net/http"
-)
-
-const (
-	authHeaderKey = "Authorization"
-
-	infoSeverity    = "INFO"
-	noticeSeverity  = "NOTICE"
-	warningSeverity = "WARNING"
-	errorSeverity   = "ERROR"
-	fatalSeverity   = "FATAL"
-
-	exceptionCategory    = "EXCEPTION"
-	alertCategory        = "ALERT"
-	notificationCategory = "NOTIFICATION"
 )
 
 // ANS holds the setup for the xsuaa service to retrieve a bearer token for authorization and
@@ -45,42 +30,11 @@ type ServiceKey struct {
 	OauthUrl     string `json:"oauth_url"`
 }
 
-// Event structure of the SAP Alert Notification Service
-type Event struct {
-	EventType      string                 `json:"eventType,omitempty"`
-	EventTimestamp int64                  `json:"eventTimestamp,omitempty"`
-	Severity       string                 `json:"severity,omitempty"`
-	Category       string                 `json:"category,omitempty"`
-	Subject        string                 `json:"subject,omitempty"`
-	Body           string                 `json:"body,omitempty"`
-	Priority       int                    `json:"priority,omitempty"`
-	Tags           map[string]interface{} `json:"tags,omitempty"`
-	Resource       *Resource              `json:"resource,omitempty"`
-}
-
-// Resource structure of the SAP Alert Notification Service Event
-type Resource struct {
-	ResourceName     string                 `json:"resourceName,omitempty"`
-	ResourceType     string                 `json:"resourceType,omitempty"`
-	ResourceInstance string                 `json:"resourceInstance,omitempty"`
-	Tags             map[string]interface{} `json:"tags,omitempty"`
-}
-
 // UnmarshallServiceKeyJSON unmarshalls the given json service key string.
 func UnmarshallServiceKeyJSON(serviceKeyJSON string) (ansServiceKey ServiceKey, err error) {
 	err = json.Unmarshal([]byte(serviceKeyJSON), &ansServiceKey)
 	if err != nil {
 		err = errors.Wrap(err, "error unmarshalling ANS serviceKey")
-		return
-	}
-	return
-}
-
-// MergeWithJSON unmarshalls an ANS Event JSON string and merges it with the existing receiver Event
-func (event *Event) MergeWithJSON(eventJSON []byte) (err error) {
-	err = json.Unmarshal(eventJSON, &event)
-	if err != nil {
-		err = errors.Wrapf(err, "error unmarshalling ANS event from JSON string %q", eventJSON)
 		return
 	}
 	return
@@ -127,31 +81,4 @@ func (ans ANS) Send(event Event) error {
 	}
 
 	return nil
-}
-
-// TranslateLogrusLogLevel takes the logrus log level and translates it to an ANS severity ans category string
-func TranslateLogrusLogLevel(level logrus.Level) (severity, category string) {
-	severity = infoSeverity
-	category = notificationCategory
-	switch level {
-	case logrus.InfoLevel:
-		severity = infoSeverity
-		category = notificationCategory
-	case logrus.DebugLevel:
-		severity = infoSeverity
-		category = notificationCategory
-	case logrus.WarnLevel:
-		severity = warningSeverity
-		category = alertCategory
-	case logrus.ErrorLevel:
-		severity = errorSeverity
-		category = exceptionCategory
-	case logrus.FatalLevel:
-		severity = fatalSeverity
-		category = exceptionCategory
-	case logrus.PanicLevel:
-		severity = fatalSeverity
-		category = exceptionCategory
-	}
-	return
 }
